格式化代码

+++ Java/javanewfeature/src/main/java/com/lee/java8/NumericStream.java
@@ -12,7 +12,7 @@ public class NumericStream {
 
   public static void main(String[] args) {
     Stream<Integer> stream = Arrays.stream(new Integer[]{1, 2, 3, 4, 5, 6, 7});
-    IntStream intStream = stream.mapToInt(i -> i.intValue());
+    IntStream intStream = stream.mapToInt(i -> i);
     int result = intStream.filter(i -> i > 3).sum();
     System.out.println(result);
